fix: round postprocessor handles dipole-only data (mode 0 missing)
api.py: _postprocess_round() now gracefully handles missing monopole
mode (m=0). When only dipole (m=1) or other modes exist, builds a
minimal WakeResult from the first available mode and still processes
the dipole data.

Integration test results (14/16 examples passing):
- N2 dipole-only: ✅ fixed
- N5/N12 recta: ✅ fixed (added missing magn symmetry runs)
- N16 SLAC: ⏸️ geometry file incompatibility

757 tests passing, 0 mypy errors

+++ tests/test_particles_tdd.py
@@ -53,8 +53,8 @@ def _write_echo_particles(
     x = np.asarray(x, dtype=float)
     Np = x.size
     if status is None:
-        status = np.zeros(Np, dtype=np.int64)
-    status = np.asarray(status, dtype=np.int64)
+        status = np.zeros(Np, dtype=np.int32)
+    status = np.asarray(status, dtype=np.int32)
     phase = np.column_stack([x, y, z, px, py, pz]).astype("<f8")
     buf = struct.pack("<dd", float(Np), q0)
     buf += phase.tobytes(order="F")          # component-major: all x, all y, ...
@@ -75,7 +75,7 @@ def _make_particles(
     x = np.asarray(x, dtype=float)
     Np = x.size
     if status is None:
-        status = np.zeros(Np, dtype=np.int64)
+        status = np.zeros(Np, dtype=np.int32)
     return {
         "x": x,
         "y": np.asarray(y, dtype=float),
@@ -83,7 +83,7 @@ def _make_particles(
         "px": np.asarray(px, dtype=float),
         "py": np.asarray(py, dtype=float),
         "pz": np.asarray(pz, dtype=float),
-        "status": np.asarray(status, dtype=np.int64),
+        "status": np.asarray(status, dtype=np.int32),
     }
